+++ src/types.ts
@@ -1,0 +1,2 @@
+export type TouchEventCallback = (event: TouchEvent, index: number) => void
+export type MouseEventCallback = (event: MouseEvent, index: number) => void